workingonBasket

+++ AppEntities/OrderAggregate/Order.cs
@@ -12,7 +12,7 @@ namespace BabALSaray.AppEntities.OrderAggregate
         }
 
         public Order(IReadOnlyList<OrderItem> orderItems,string buyerEmail, OrderAddress orderAddress,
-         OrderMethod orderMethod,  double subtotal)
+         OrderMethod orderMethod,  decimal subtotal)
         {
             BuyerEmail = buyerEmail;
             OrderAddress = orderAddress;
@@ -27,13 +27,13 @@ namespace BabALSaray.AppEntities.OrderAggregate
         public OrderAddress OrderAddress { get; set; }
         public OrderMethod  OrderMethod { get; set; }
         public IReadOnlyList<OrderItem> OrderItems { get; set; }
-        public double Subtotal { get; set; }
+        public decimal Subtotal { get; set; }
         
         [Column(TypeName = "nvarchar(10)")]
         public OrderStatus Status { get; set; } = OrderStatus.Pending;
         public string PaymentIntentId { get; set; }
 
-        public double GetTotal()
+        public decimal GetTotal()
         {
            return Subtotal + OrderMethod.Price;
         }
